server/requirements.txt
Flask==3.0.0
flask-cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.0.0
requests==2.32.3
groq==0.32.0
azure-ai-vision-imageanalysis==1.0.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-ai-formrecognizer==3.3.3
azure-storage-blob==12.19.0
msrest==0.7.1